add new file for benchmark confirmer
